tests/requirements.txt
django-crispy-forms>=1.4.0
django-autoslug
flake8
coverage
SQLAlchemy
pytest-django
pytest-factoryboy
pytest-cov
ipdb
